requirements.txt
Flask>=3.0.0
Flask-Session>=0.8.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.111.0
